also use takeEveryUnlessBusy for copySegmentationLayer
