+++ frontend/javascripts/oxalis/model/sagas/volumetracing_saga.js
@@ -4,15 +4,8 @@ import { message } from "antd";
 import React from "react";
 import _ from "lodash";
 
+import type { Action } from "oxalis/model/actions/actions";
 import { CONTOUR_COLOR_DELETE, CONTOUR_COLOR_NORMAL } from "oxalis/geometries/contourgeometry";
-import {
-  type CopySegmentationLayerAction,
-  updateDirectionAction,
-  updateSegmentAction,
-  finishAnnotationStrokeAction,
-  type SetActiveCellAction,
-  type ClickSegmentAction,
-} from "oxalis/model/actions/volumetracing_actions";
 import {
   ResolutionInfo,
   getBoundaries,
@@ -69,6 +62,14 @@ import {
 } from "oxalis/model/accessors/tool_accessor";
 import { markVolumeTransactionEnd } from "oxalis/model/bucket_data_handling/bucket";
 import { setToolAction, setBusyBlockingInfoAction } from "oxalis/model/actions/ui_actions";
+import { takeEveryUnlessBusy } from "oxalis/model/sagas/saga_helpers";
+import {
+  updateDirectionAction,
+  updateSegmentAction,
+  finishAnnotationStrokeAction,
+  type SetActiveCellAction,
+  type ClickSegmentAction,
+} from "oxalis/model/actions/volumetracing_actions";
 import {
   updateTemporarySettingAction,
   type UpdateTemporarySettingAction,
@@ -109,7 +110,11 @@ import sampleVoxelMapToResolution, {
 
 export function* watchVolumeTracingAsync(): Saga<void> {
   yield* take("WK_READY");
-  yield _takeEvery("COPY_SEGMENTATION_LAYER", copySegmentationLayer);
+  yield* takeEveryUnlessBusy(
+    "COPY_SEGMENTATION_LAYER",
+    copySegmentationLayer,
+    "Copying from neighbor slice",
+  );
   yield _takeLeading("INFER_SEGMENT_IN_VIEWPORT", inferSegmentInViewport);
   yield* fork(warnOfTooLowOpacity);
 }
@@ -424,7 +429,10 @@ function* labelWithVoxelBuffer2D(
   );
 }
 
-function* copySegmentationLayer(action: CopySegmentationLayerAction): Saga<void> {
+function* copySegmentationLayer(action: Action): Saga<void> {
+  if (action.type !== "COPY_SEGMENTATION_LAYER") {
+    throw new Error("Satisfy flow");
+  }
   const allowUpdate = yield* select(state => state.tracing.restrictions.allowUpdate);
   if (!allowUpdate) return;
